requirements-dev.txt
-r requirements.txt

flake8
pytest

django-extensions==3.2.1
pytest-django==4.9.0
